Undo moving functions into utility class
